pyproject.toml
[project]
name = "project-chimera"
version = "0.1.0"
description = "Autonomous AI Influencer Factory"
authors = [{name = "Bethelhem Abay"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pydantic-settings>=2.1.0",
    "tenacity>=8.2.0",
    "redis>=5.0.0",
    "weaviate-client>=3.25.0",
    "mcp>=0.9.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-mock", "black", "mypy", "ruff"]
ml = ["openai", "langchain", "chromadb"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
python-preference = "only-managed"